Bug: Fixed the problem with currency exchange rate returning data in string

+++ backend/database/migrations/20240831115710-create-currency-exchange-rates.js
@@ -12,19 +12,19 @@ module.exports = {
 
       
       dolarBCVToBs: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
         allowNull: false,
       },
       eurosBCVToBs: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
         allowNull: false,
       },
       dolarBlackToBs: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
         // allowNull: false,
       },
       euroBlackToBs: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL,
         // allowNull: false,
       },
